requirements.txt
pytesseract==0.3.3
pandas==1.0.3
Pillow==7.2.0
